refactor: Move all request/response structs to dedicated dto package
