+++ pkg/handlers/vulnerability.go
@@ -10,6 +10,7 @@ import (
 	"github.com/kptm-tools/core-service/pkg/api"
 	"github.com/kptm-tools/core-service/pkg/customerrors"
 	"github.com/kptm-tools/core-service/pkg/interfaces"
+	"github.com/kptm-tools/core-service/pkg/ws/report/dto"
 )
 
 type VulnerabilityHandlers struct {
@@ -48,16 +49,16 @@ func (h *VulnerabilityHandlers) GetVulnerability(w http.ResponseWriter, r *http.
 		})
 	}
 
-	vulnDetailResponse := ScanVulnerabilityDetailResponse{
+	vulnDetailResponse := dto.ScanVulnerabilityDetailResponse{
 		ID:       vulnDetail.ID,
 		ScanDate: vulnDetail.ScanDate,
 
-		Host: HostItem{
+		Host: dto.HostItem{
 			Alias:     vulnDetail.Host.Alias,
 			IPAddress: vulnDetail.Host.IPAddress,
 		},
-		Port: adaptDomainPortItem(vulnDetail.Port),
-		OS:   adaptDomainOSItem(vulnDetail.OS),
+		Port: dto.AdaptDomainPortItem(vulnDetail.Port),
+		OS:   dto.AdaptDomainOSItem(vulnDetail.OS),
 
 		Name:           vulnDetail.Name,
 		Severity:       vulnDetail.Severity,
@@ -75,10 +76,10 @@ func (h *VulnerabilityHandlers) GetVulnerability(w http.ResponseWriter, r *http.
 		VendorComments: vulnDetail.VendorComments,
 		References:     vulnDetail.References,
 
-		DateInfo:   adaptDomainDateInfo(vulnDetail.DateInfo),
-		PluginInfo: adaptDomainPluginInfo(vulnDetail.PluginInfo),
-		VPRKeyD:    adaptDomainVPRKeyInfo(vulnDetail.VPRKeyD),
-		RiskInfo:   adaptDomainRiskInfo(vulnDetail.RiskInfo),
+		DateInfo:   dto.AdaptDomainDateInfo(vulnDetail.DateInfo),
+		PluginInfo: dto.AdaptDomainPluginInfo(vulnDetail.PluginInfo),
+		VPRKeyD:    dto.AdaptDomainVPRKeyInfo(vulnDetail.VPRKeyD),
+		RiskInfo:   dto.AdaptDomainRiskInfo(vulnDetail.RiskInfo),
 	}
 
 	return api.WriteJSON(w, http.StatusOK, vulnDetailResponse)
@@ -91,7 +92,7 @@ func (h *VulnerabilityHandlers) CreateVulnerabilityComment(w http.ResponseWriter
 			Error: fmt.Sprintf("ID is invalid: %v", err),
 		})
 	}
-	vulnCommentRequest := new(VulnerabilityCommentRequest)
+	vulnCommentRequest := new(dto.VulnerabilityCommentRequest)
 	if err := decodeJSONBody(w, r, vulnCommentRequest); err != nil {
 		var mr *malformedRequest
 
@@ -123,7 +124,7 @@ func (h *VulnerabilityHandlers) PatchVulnerabilityComment(w http.ResponseWriter,
 			Error: fmt.Sprintf("ID is invalid: %v", err),
 		})
 	}
-	vulnCommentRequest := new(VulnerabilityCommentRequest)
+	vulnCommentRequest := new(dto.VulnerabilityCommentRequest)
 	if err := decodeJSONBody(w, r, vulnCommentRequest); err != nil {
 		var mr *malformedRequest
